# packages/shared-types/schemas/openapi.json
{
  "components": {
    "schemas": {
      "AccessPoint": {
        "additionalProperties": false,
        "description": "Normalized wireless access point snapshot.",
        "properties": {
          "band": {
            "anyOf": [
              {
                "maxLength": 16,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Band"
          },
          "bssid": {
            "pattern": "^[0-9A-Fa-f]{2}(:[0-9A-Fa-f]{2}){5}$",
            "title": "Bssid",
            "type": "string"
          },
          "channel": {
            "anyOf": [
              {
                "maximum": 196.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "encryption": {
            "items": {
              "type": "string"
            },
            "title": "Encryption",
            "type": "array"
          },
          "first_seen": {
            "format": "date-time",
            "title": "First Seen",
            "type": "string"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "title": "Flags",
            "type": "array"
          },
          "last_seen": {
            "format": "date-time",
            "title": "Last Seen",
            "type": "string"
          },
          "signal_history": {
            "items": {
              "$ref": "#/components/schemas/SignalSample"
            },
            "title": "Signal History",
            "type": "array"
          },
          "ssid": {
            "default": "",
            "maxLength": 128,
            "title": "Ssid",
            "type": "string"
          },
          "vendor_oui": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Oui"
          }
        },
        "required": [
          "bssid"
        ],
        "title": "AccessPoint",
        "type": "object"
      },
...
      "AllowTargetRequest": {
        "description": "Allow-list target request.",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/TargetKind"
          },
          "value": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "value"
        ],
        "title": "AllowTargetRequest",
        "type": "object"
      },
      "ApiPage_AccessPoint_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AccessPoint"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ApiPage[AccessPoint]",
        "type": "object"
      },
...
      "ApiPage_Client_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Client"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ApiPage[Client]",
        "type": "object"
      },
      "ApiPage_Event_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ApiPage[Event]",
        "type": "object"
      },
...
      "Client": {
        "additionalProperties": false,
        "description": "Normalized wireless client snapshot.",
        "properties": {
          "associated_bssid": {
            "anyOf": [
              {
                "pattern": "^[0-9A-Fa-f]{2}(:[0-9A-Fa-f]{2}){5}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Associated Bssid"
          },
          "first_seen": {
            "format": "date-time",
            "title": "First Seen",
            "type": "string"
          },
          "last_seen": {
            "format": "date-time",
            "title": "Last Seen",
            "type": "string"
          },
          "mac": {
            "pattern": "^[0-9A-Fa-f]{2}(:[0-9A-Fa-f]{2}){5}$",
            "title": "Mac",
            "type": "string"
          },
          "probes": {
            "items": {
              "type": "string"
            },
            "title": "Probes",
            "type": "array"
          },
          "signal_history": {
            "items": {
              "$ref": "#/components/schemas/SignalSample"
            },
            "title": "Signal History",
            "type": "array"
          },
          "vendor_oui": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Oui"
          }
        },
        "required": [
          "mac"
        ],
        "title": "Client",
        "type": "object"
      },
      "Engagement": {
        "additionalProperties": false,
        "description": "Engagement scope and lifecycle metadata.",
        "properties": {
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "scope_rules": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Scope Rules",
            "type": "array"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "Engagement",
        "type": "object"
      },
      "EngagementCreateRequest": {
        "description": "Engagement creation request.",
        "properties": {
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "scope_rules": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Scope Rules",
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "title": "EngagementCreateRequest",
        "type": "object"
      },
      "Event": {
        "additionalProperties": false,
        "description": "Append-only event record.",
        "properties": {
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/EventKind"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "sensor_id": {
            "maxLength": 96,
            "minLength": 1,
            "title": "Sensor Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sensor_id",
          "kind",
          "payload"
        ],
        "title": "Event",
        "type": "object"
      },
      "EventKind": {
        "description": "Normalized event kinds emitted by sensors.",
        "enum": [
          "access_point_seen",
          "client_seen",
          "sensor_status",
          "command_result"
        ],
        "title": "EventKind",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SignalSample": {
        "additionalProperties": false,
        "description": "Signal reading at a point in time.",
        "properties": {
          "rssi_dbm": {
            "maximum": 20.0,
            "minimum": -127.0,
            "title": "Rssi Dbm",
            "type": "integer"
          },
          "seen_at": {
            "format": "date-time",
            "title": "Seen At",
            "type": "string"
          }
        },
        "required": [
          "rssi_dbm"
        ],
        "title": "SignalSample",
        "type": "object"
      },
...
      "TargetKind": {
        "description": "Target identifier kinds used by active-operation allow-lists.",
        "enum": [
          "bssid",
          "ssid",
          "client_mac"
        ],
        "title": "TargetKind",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Cheeky Pony API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/access_points": {
      "get": {
        "description": "List observed access points.\n\nArgs:\n    _: Current user.\n    store: Application store.\n    limit: Page size.\n    offset: Page offset.\n\nReturns:\n    Paginated access point list.",
        "operationId": "list_access_points_api_v1_access_points_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiPage_AccessPoint_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Access Points",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/access_points/{bssid}": {
      "get": {
        "description": "Return one access point.\n\nArgs:\n    bssid: BSSID.\n    _: Current user.\n    store: Application store.\n\nReturns:\n    Access point.",
        "operationId": "get_access_point_api_v1_access_points__bssid__get",
        "parameters": [
          {
            "in": "path",
            "name": "bssid",
            "required": true,
            "schema": {
              "title": "Bssid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessPoint"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Access Point",
        "tags": [
          "devices"
        ]
      }
    },
...
    "/api/v1/devices": {
      "get": {
        "description": "List observed client devices.\n\nArgs:\n    _: Current user.\n    store: Application store.\n    limit: Page size.\n    offset: Page offset.\n\nReturns:\n    Paginated client list.",
        "operationId": "list_clients_api_v1_devices_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiPage_Client_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Clients",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{mac}": {
      "get": {
        "description": "Return one client device.\n\nArgs:\n    mac: Client MAC.\n    _: Current user.\n    store: Application store.\n\nReturns:\n    Client device.",
        "operationId": "get_client_api_v1_devices__mac__get",
        "parameters": [
          {
            "in": "path",
            "name": "mac",
            "required": true,
            "schema": {
              "title": "Mac",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Client",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/engagements": {
      "post": {
        "description": "Create an engagement.\n\nArgs:\n    payload: Engagement payload.\n    _: Current admin with verified TOTP.\n    store: Application store.\n\nReturns:\n    Created engagement.",
        "operationId": "create_engagement_api_v1_engagements_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Engagement"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Engagement",
        "tags": [
          "engagements"
        ]
      }
    },
    "/api/v1/engagements/{engagement_id}/allow-list": {
      "post": {
        "description": "Add a target to an engagement allow-list.\n\nArgs:\n    engagement_id: Engagement identifier.\n    payload: Target payload.\n    _: Current admin with verified TOTP.\n    store: Application store.",
        "operationId": "allow_target_api_v1_engagements__engagement_id__allow_list_post",
        "parameters": [
          {
            "in": "path",
            "name": "engagement_id",
            "required": true,
            "schema": {
              "title": "Engagement Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllowTargetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Allow Target",
        "tags": [
          "engagements"
        ]
      }
    },
    "/api/v1/events": {
      "get": {
        "description": "List events.\n\nArgs:\n    _: Current user.\n    store: Application store.\n    limit: Page size.\n    offset: Page offset.\n\nReturns:\n    Paginated event list.",
        "operationId": "list_events_api_v1_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiPage_Event_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Events",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/events/{event_id}": {
      "get": {
        "description": "Return one event.\n\nArgs:\n    event_id: Event identifier.\n    _: Current user.\n    store: Application store.\n\nReturns:\n    Event.",
        "operationId": "get_event_api_v1_events__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Event",
        "tags": [
          "devices"
        ]
      }
    },
...
  }
}